mountlib: refactor before adding serve docker (#5415)
